h_pulse/requirements.txt
python>=3.10
numpy>=1.24
scipy>=1.10
pandas>=2.0
pydantic>=2.5
matplotlib>=3.7
plotly>=5.15
fastapi>=0.110
uvicorn>=0.23
orjson>=3.9
httpx>=0.24
jinja2>=3.1
reportlab>=4.0
rich>=13.6
typer>=0.9
structlog>=23.1
cryptography>=41.0
pysha3>=1.0
web3>=6.11
python-dateutil>=2.8
pytz>=2023.3
astropy>=5.3
skyfield>=1.46
jplephem>=2.22
pyswisseph>=2.10.3.2
astral>=3.2
torch>=2.2; platform_system!="Darwin" or platform_machine!="arm64"
transformers>=4.39
networkx>=3.1
qiskit>=1.0
pennylane>=0.35
uvloop>=0.19
strawberry-graphql>=0.211
pytest>=7.4
starlette>=0.36